Set sides of Second Tile

+++ app/src/main/java/com/delek/lostrealm/database/model/Map.kt
@@ -5,6 +5,6 @@ data class Map(
     val tileId: Int,
     val posX: Int,
     val posY: Int,
-    val rotate: Float,
+    val rotate: Int,
     val enchant: Int
 )
